добавлена функция проверки введенных пользователем запрещенных символов

+++ src/main.py
@@ -6,7 +6,9 @@ import time
 from dotenv import load_dotenv
 import telebot
 
-from constants import GREETING_TEXT, HELP_TEXT, TYPE_GAME_NAME, WAITING_TEXT
+from constants import (
+    GREETING_TEXT, HELP_TEXT, TYPE_GAME_NAME, WAITING_TEXT, BAN_SYMBOLS,
+    KEYS_NOT_FOUND)
 from functions import platimarket, steampay, print_result, games_free
 
 time.sleep(5)
@@ -18,54 +20,58 @@ bot = telebot.TeleBot(os.getenv('TOKEN'))
 @bot.message_handler(commands=['start'])
 def start(message: telebot.types.Message) -> telebot.types.Message:
     """Выводит сообщение при входе"""
-
     return bot.send_message(message.chat.id, GREETING_TEXT, parse_mode='html')
 
 
 @bot.message_handler(commands=['help'])
 def helper(message: telebot.types.Message) -> telebot.types.Message:
     """Показывает основные команды"""
-
     return bot.send_message(message.chat.id, HELP_TEXT, parse_mode='html')
 
 
 @bot.message_handler(commands=['search'])
 def search_game(message: telebot.types.Message) -> None:
     """Выводит сообщение, после которого запускается функция поиска ключей"""
-
     bot.send_message(
         message.chat.id, TYPE_GAME_NAME)
-
     bot.register_next_step_handler(message, find_keys)
 
 
 def find_keys(message: telebot.types.Message) -> None:
     """Находит самые дешевые ключи игр"""
 
-    msg = bot.send_message(message.chat.id, WAITING_TEXT)
-
     game_name = message.text.lower()
 
-    # словарь для добавления цены, ссылки и названий игр
-    dict_price_url = {}
+    # проверка на введенные пользователем запрещеные символы
+    bad_sym = check_bad_symbols(game_name, message)
 
-    # Поиск ключей на сайте plati.ru
-    platimarket.plati(game_name, dict_price_url)
+    if not bad_sym:
 
-    # Поиск ключей на сайте steampay.com
-    steampay.steam_pay(game_name, dict_price_url)
+        msg = bot.send_message(message.chat.id, WAITING_TEXT)
 
-    # Сортировка полученных цен по возрастанию
-    sorted_prices = sorted(dict_price_url.items())
+        # словарь для добавления цены, ссылки и названий игр
+        dict_price_url = {}
 
-    # Удалить сообщение msg
-    bot.delete_message(message.chat.id, msg.id)
+        # Поиск ключей на сайте plati.ru
+        platimarket.plati(game_name, dict_price_url)
 
-    # Вывод результата для топ-5 цен
-    print_result.print_result(sorted_prices, bot, message)
+        # Поиск ключей на сайте steampay.com
+        steampay.steam_pay(game_name, dict_price_url)
 
-    # Что делать после вывода результата
-    bot.register_next_step_handler(message, step_after_find_keys)
+        # Сортировка полученных цен по возрастанию
+        sorted_prices = sorted(dict_price_url.items())
+
+        # Удалить сообщение msg
+        bot.delete_message(message.chat.id, msg.id)
+
+        # Вывод результата для топ-5 цен
+        print_result.print_result(sorted_prices, bot, message)
+
+        # Что делать после вывода результата
+        bot.register_next_step_handler(message, step_after_find_keys)
+
+    else:
+        bot.register_next_step_handler(message, find_keys)
 
 
 def step_after_find_keys(message: telebot.types.Message) -> None:
@@ -85,12 +91,19 @@ def step_after_find_keys(message: telebot.types.Message) -> None:
         find_keys(message)
 
 
+def check_bad_symbols(game_name: str, message: telebot.types.Message) \
+        -> bool:
+    """Не дает искать название игры из сообщения с запрещенными символами"""
+    if set(game_name) & BAN_SYMBOLS:
+        bot.send_message(message.chat.id, KEYS_NOT_FOUND)
+        return True
+
+
 @bot.message_handler(commands=['free'])
 def search_free_games(message: telebot.types.Message) -> None:
     """Находит раздачи бесплатных игр"""
 
-    mseg = bot.send_message(
-        message.chat.id, WAITING_TEXT)
+    mseg = bot.send_message(message.chat.id, WAITING_TEXT)
 
     # Получить информацию о раздаче игр
     games_free.free_games(bot, message)
@@ -116,4 +129,5 @@ def step_after_search_free_games(message: telebot.types.Message) -> None:
         helper(message)
 
 
-bot.polling(none_stop=True)
+if __name__ == '__main__':
+    bot.polling(none_stop=True)
